Breadcrumb component's navigation export to BreadcrumNavigation component
